Rename accordion stories
