+++ .eslintrc.json
@@ -64,7 +64,7 @@
       }
     },
     {
-      "files": ["*.spec.{js,ts,tsx}", "*.stories.*", "utils/**/*.{js,ts,tsx}", ".storybook/**/*.{js,ts,tsx}", "styleguide.config.js"],
+      "files": ["*.spec.{js,ts,tsx}", "**/stories/**/*", "utils/**/*.{js,ts,tsx}", ".storybook/**/*.{js,ts,tsx}", "styleguide.config.js"],
       "rules": {
         "garden-local/require-default-theme": "off",
         "@typescript-eslint/no-var-requires": "off"
